requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMuPDF==1.23.8
python-docx==0.8.11
openai==0.28.1
plotly==5.17.0
pandas==2.1.3
reportlab==4.0.7
Jinja2==3.1.2
Werkzeug==2.3.7
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
gunicorn==21.2.0
reportlab==4.0.7